Dev to login

+++ app.py
@@ -30,12 +30,12 @@ import hashlib
 
 import random
 
+
 # HTML을 주는 부분
 @app.route('/')
 def home():
-
-
     return render_template('index.html')
+
 
 @app.route('/login')
 def login():
@@ -47,17 +47,12 @@ def register():
     return render_template('register.html')
 
 
-@app.route('/mypage')
-def mypage():
-    return render_template('mypage.html')
-
-
 @app.route('/exhibition', methods=['GET'])
 def exhibition():
     with conn.cursor() as cursor:
         cntSql = "SELECT count(*), exhibition_id FROM exhibitions GROUP BY exhibition_id"
         cursor.execute(cntSql)
-
+        today = datetime.datetime.now()
         cnt = cursor.fetchall()
         rowCount = len(cnt)
         idList = []
@@ -79,8 +74,8 @@ def exhibition():
             randNumList.append(idList[randNum])
 
         for j in range(0, cardNum):
-            selectSql = "SELECT * FROM exhibitions WHERE exhibition_id = %s"
-            cursor.execute(selectSql, (randNumList[j]))
+            selectSql = "SELECT * FROM exhibitions WHERE exhibition_id = %s AND end_date > %s"
+            cursor.execute(selectSql, (randNumList[j], today))
             result = cursor.fetchall()
             resultList.append(result)
 
@@ -100,13 +95,13 @@ def like():
         if cnt <= 0:
             return jsonify({'msg': '찜목록 추가 실패!'})
         else:
-            dupCheckSql = "SELECT count(*) FROM WishList WHERE exhibition_id2 = %s"
-            cursor.execute(dupCheckSql, int(exhibitionId_receive))
+            dupCheckSql = "SELECT count(*) FROM WishList WHERE exhibition_id2 = %s AND user_id2 = %s"
+            cursor.execute(dupCheckSql, (int(exhibitionId_receive), userId_receive))
             cnt2 = cursor.fetchall()
 
             if cnt2[0][0] != 0:
-                deleteSql = "DELETE FROM WishList WHERE exhibition_id2 = %s"
-                cursor.execute(deleteSql, int(exhibitionId_receive))
+                deleteSql = "DELETE FROM WishList WHERE exhibition_id2 = %s AND user_id2 = %s"
+                cursor.execute(deleteSql, (int(exhibitionId_receive), userId_receive))
                 conn.commit()
                 return jsonify({'msg': '찜목록에서 제거되었습니다!'})
             else:
@@ -115,18 +110,6 @@ def like():
                 print(result)
                 conn.commit()
                 return jsonify({'msg': '찜목록에 추가되었습니다!'})
-
-@app.route('/user_like', methods=['GET'])
-def userLike():
-    userId_receive = request.args.get('userId')
-    print(userId_receive)
-    with conn.cursor() as cursor:
-        sql = "SELECT exhibition_id2 FROM WishList WHERE user_id2 = %s"
-        cursor.execute(sql, (userId_receive))
-        result = cursor.fetchall()
-        print(result)
-    return jsonify({'results': result})
-
 
 # 로그인, 회원가입을 위한 API
 
@@ -207,7 +190,6 @@ def getDetail(id):
     getId = cursor.fetchone();
     print(getId)
 
-
     if not getId: return render_template("index.html")
 
     getExhibitionSql = "select * from exhibitions where exhibition_id = %s"
@@ -278,6 +260,40 @@ def deleteReview(id):
         conn.commit()
 
     return jsonify({'msg': '삭제 완료!'})
+
+
+@app.route('/mypage')
+def mypage():
+
+    tokenReceive = request.cookies.get('mytoken')
+
+    payload = jwt.decode(tokenReceive, SECRET_KEY, algorithms=['HS256'])
+    # userId를 DB에서 찾는다.
+    print(payload)
+
+    sql = "SELECT exhibition_id,init_date, end_date, title, img_url, place " \
+              "FROM mini.wishlist RIGHT JOIN mini.exhibitions "\
+              "on wishlist.exhibition_id2 = exhibitions.exhibition_id " \
+              "where user_id2 = %s"
+
+
+    cursor.execute(sql, (payload['userId']))
+
+    print(cursor)
+
+    likes = cursor.fetchall()
+
+    print(likes)
+
+    getReviewSql = "select e.exhibition_id, e.title, e.img_url, r.content from reviews as r left join exhibitions as e on r.exhibition_id2 = e.exhibition_id where r.user_id2 = %s"
+
+
+    cursor.execute(getReviewSql, (payload['userId']))
+
+    reviews = cursor.fetchall()
+    print(reviews)
+
+    return render_template('mypage.html', likes=likes, reviews=reviews)
 
 
 if __name__ == '__main__':
